Fix relative paths failing to resolve in diffs
The previous implementation used git difftool to automatically run a
script against the files that had changed. git difftool automatically
created copies of the files being diffed in a temporary location, but
did not copy rest of the project, causing relative paths to no longer
resolve.

This change removes the use of git difftool, and instead exports a copy
of the repo at the pull request target before invoking the diff
generator on the changed VIs. This allows relative paths to be resolved
correctly, leading to nicer diffs.

This change also enables generating diffs for .vit and .vim files.