Diff integration (#13)
* Update diffPipeline.groovy

* Update diffPipeline.groovy

* Adding Diff capabilities

This update adds the dependancies for the Diff CLI step to the repository.

* Making the nodeLabel optional

The nodeLabel is now effectively optional. If there is no nodeLabel passed to the executor, it will use an empty string. On an empty string, it will run it on the node with the correct LabVIEW version.

* Update Pipeline.groovy

* Update PipelineInformation.groovy

Changing the echo to more accurately print out the node.

* Update PipelineInformation.groovy

* Update PipelineInformation.groovy

* Changing to using trim() to handle null

From buckd's comment, the previous version would not work correctly if pipelineInformation was null. Changing to using trim() instead.

* Update Pipeline.groovy

* Adding single screenshot VI

* UI Updates to Screenshot Single VI

Moved items to the top left for easier single-screen access.

* UI Updates

Moved the majority of the VIs to be saved on the top left of the screen.

Moved some controls that were overlapping.

* Changing Nomenclature from VI 1 and 2 to Old and New respectively

* Fixing the single VI case

Was passing the incorrect VI to the single VI case. Also forgot to add the upload path for that case.

* Adding the exe

* Fixing titles on the images

The Diff images contained duplicated VI names.

* Update ScreenshotSingleVIs.vi

Added suffix to the image text for the single VI case.

* Style changes

* Update Pipeline.groovy

Style changes.

* Update PipelineExecutor.groovy

Style changes (line after } for consistency)

* Update PipelineInformation.groovy

* Update PipelineExecutor.groovy

* Update Pipeline.groovy

* Update readme.md

Formatting and phrasing changes.

* PR Update for separating source and some style changes

Code Organization:
- Moved code to main project, removed Diff specific project
- Reorganized code to be within the appropriate libraries
- Ensured all code was separate from source
Style Changes
- Homogenized naming of Diff VIs
- Wrapped VIs in case structures for the error-in case for homogeneity.
Documentation
- Adding mention that the OldVI parameter is optional
- Spelling changes
- Changed readme.md to uppercase (using CLI)

* PR Changes

- gitignore changes
Removed the Executables header and the .dll line to allow for uploading a necessary dependent dll.
- RandomString
The current scheme is somewhat convoluted. Changed it to something slightly simpler and added a comment
- Diff VIs
Style changes