Moving build.toml help from Wiki to file in docs folder (#20)
* Moving nodeLabel to JenkinsFile

Pipeline Executor now expects the first argument to be the Label of the node.

This would allow for building arbitrary, non-veristand projects with this system.

* Update diffPipeline.groovy

* Update diffPipeline.groovy

* Making the nodeLabel optional

The nodeLabel is now effectively optional. If there is no nodeLabel passed to the executor, it will use an empty string. On an empty string, it will run it on the node with the correct LabVIEW version.

* Update Pipeline.groovy

* Update PipelineInformation.groovy

Changing the echo to more accurately print out the node.

* Update PipelineInformation.groovy

* Update PipelineInformation.groovy

* Changing to using trim() to handle null

From buckd's comment, the previous version would not work correctly if pipelineInformation was null. Changing to using trim() instead.

* Update Pipeline.groovy

* Update Pipeline.groovy

Style changes.

* Update PipelineExecutor.groovy

Style changes (line after } for consistency)

* Update PipelineInformation.groovy

* Update PipelineExecutor.groovy

* Update Pipeline.groovy

* Moving Wiki on TOML attributes to a Markdown file on a new docs folder.

Moving Wiki on TOML attributes to a Markdown file on a new docs folder.

Changing Link on the main README.

* Update README.md

Forgot file extension from relative link.

* Update README.md

Escaping space.

* Update Toml Help.md

Changing TOC to use Anchors.

* Update Toml Help.md

Noticed that a link to the readme pointed to a readme in an outdated repo.

* Update Toml Help.md

Found another link to the commonbuild repo.