Build packages for pull requests that modify package-related files (#86)
Allow package steps to specify files that impact their output. This is used as a heuristic for rebuilding packages during pull requests. Currently this includes the build.toml file, and control and instructions files for nipkg steps.

We currently either have to manually force packages to be built when testing, or blindly submit code and wait for a release branch to pick it up. This should give us more confidence when reviewing changes, without building packages for every change.